refactor: Moved types out of ColorizeCanvas
